使用宏进行签名函数展开，避免函数指针类型强转以及对编译器传参方式的依赖。由于新函数签名不使用没有额外开销，故移除是否使用函数签名的选项。函数代理函数因为依赖于编译器传参方式，也一并移除。此外，函数说明改为字符串，便于语法高亮识别。

+++ extensions/shell_enhance/shell_passthrough.h
@@ -41,7 +41,7 @@ typedef int (*ShellPassthrough)(char *data, unsigned short len);
  * @param _handler passthrough 模式 handler
  */
 #define SHELL_PASSTROUGH_FUNC(_name, _prompt, _handler) \
-        void SHELL_PASSTROUGH_FUNC_NAME(_name)(int p1, int p2) \
+        void SHELL_PASSTROUGH_FUNC_NAME(_name)(uintptr_t p1, uintptr_t p2) \
         { shellPassthrough(shellGetCurrent(), #_prompt, _handler, p1, (void *)p2); }
 
 /**
